Fix color key transparency on some systems
On a macOS machine with an Intel HD5000 iGPU with libraries installed
by nix, color key transparency does not work (i.e. the fuchsia pixels
intended to be transparent are instead visible). Changing the PNG files
to remove their alpha channels addresses the problem, but this change
also fixes the issue without changing the art resources.

It does so by copying the loaded image into a new SDL Surface
optimized for blitting to an RGBA8888 surface using the
SDL_ConvertSurface function.